+++ computers/src/main/java/dev/jstech/computers/machine/MachineServices.java
@@ -257,7 +257,7 @@ public final class MachineServices implements IHost {
             this.shell = new ServerCliComputer(terminal, server);
             this.files = new FileService(this.machine, server,
                     new NetworkPathResolver(server, this.shell::machinesNamed, this.shell::networkShares),
-                    this.shell::currentLocation, this.shell);
+                    this.shell::currentLocation);
             this.computer = new ComputerInfoService(this.machine, this.shell);
             this.network = new NetworkReadService(terminal, server, this.shell, this.shell);
             this.mainframe = new MainframeStatsService(this.shell, this.shell);